refactored trainerboard code into folders i.e components and utils inside trainerboard folder
